pyproject.toml
[tool.poetry]
name = "automated-mggg-qa"
version = "0.1.0"
description = "Automated QA checks for mggg-states"
authors = ["Max Fan <theinnovativeinventor@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
gdutils = { git = "https://github.com/InnovativeInventor/gdutils.git", branch = "main"}
pydantic = "^1.7.3"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
